Split statistics.py into modules
