+++ app/stats/statistics/__init__.py
@@ -1,0 +1,3 @@
+from .biggest_trips import biggest_trips
+from .most_common import most_common
+from .yearly import yearly
